Several improvements and API changes
This commit brigns several improvements to the code quality and
performances.

In particular, the Regex is no longer used, as per this recommendation[1].
Moreover, HTTPoison is no longer in use, since it would just clutter
the dependency tree (projects may use HTTPotion, plain Hackney or
Tesla.
The ExDoc dependency is also up-to-date.

[1]: https://stackoverflow.com/a/1732454/8913528